backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
python-dotenv
google-cloud-firestore

fastapi
pydantic
firebase-admin